IRIS-3963 | merge SignupTests and SignupTestsOld
