+++ src/test/java/com/wikia/webdriver/testcases/auth/SignupTests.java
@@ -1,16 +1,26 @@
 package com.wikia.webdriver.testcases.auth;
 
+import com.wikia.webdriver.common.contentpatterns.MercurySubpages;
+import com.wikia.webdriver.common.contentpatterns.MercuryWikis;
 import com.wikia.webdriver.common.contentpatterns.PageContent;
 import com.wikia.webdriver.common.core.annotations.Execute;
+import com.wikia.webdriver.common.core.annotations.InBrowser;
 import com.wikia.webdriver.common.core.configuration.Configuration;
+import com.wikia.webdriver.common.core.drivers.Browser;
+import com.wikia.webdriver.common.core.helpers.Emulator;
 import com.wikia.webdriver.common.properties.Credentials;
 import com.wikia.webdriver.common.templates.NewTestTemplate;
+import com.wikia.webdriver.common.users.CreateUser;
+import com.wikia.webdriver.common.users.TestUser;
+import com.wikia.webdriver.elements.common.Navigate;
+import com.wikia.webdriver.elements.mercury.old.SignupPageObject;
 import com.wikia.webdriver.pageobjectsfactory.componentobject.global_navitagtion.NavigationBar;
 import com.wikia.webdriver.pageobjectsfactory.pageobject.WikiBasePageObject;
 import com.wikia.webdriver.pageobjectsfactory.pageobject.auth.register.AttachedRegisterPage;
 
 import com.wikia.webdriver.pageobjectsfactory.pageobject.auth.register.DetachedRegisterPage;
 import com.wikia.webdriver.pageobjectsfactory.pageobject.auth.signin.DetachedSignInPage;
+import org.joda.time.DateTime;
 import org.testng.annotations.Test;
 
 import java.util.Calendar;
@@ -19,11 +29,11 @@ import static  com.wikia.webdriver.common.core.Assertion.assertEquals;
 import static  com.wikia.webdriver.common.core.Assertion.assertStringContains;
 
 @Test(groups = "auth-signUp")
+@Execute(onWikia = MercuryWikis.MERCURY_AUTOMATION_TESTING)
 public class SignupTests extends NewTestTemplate {
 
   Credentials credentials = Configuration.getCredentials();
 
-  @Test(groups = "SignUp_anonCanNotSignUpIfSheIsYoungerThanTwelve")
   public void anonCanNotSignUpIfYoungerThanTwelve() {
     WikiBasePageObject base = new WikiBasePageObject();
     base.openSpecialUserSignUpPage(wikiURL);
@@ -41,7 +51,6 @@ public class SignupTests extends NewTestTemplate {
     assertEquals(register.getError(), "We cannot complete your registration at this time");
   }
 
-  @Test(groups = "SignUp_anonCanNotSignUpIfTheUsernameAlreadyExists")
   public void anonCanNotSignUpIfTheUsernameAlreadyExists() {
     WikiBasePageObject base = new WikiBasePageObject();
     base.openSpecialUserSignUpPage(wikiURL);
@@ -59,7 +68,6 @@ public class SignupTests extends NewTestTemplate {
 
   }
 
-  @Test(groups = "SignUp_anonCanSignUpOnNewBaseAuthPageFromGlobalNav")
   public void anonCanSignUpOnNewBaseAuthPageFromGlobalNav() {
     WikiBasePageObject base = new WikiBasePageObject();
     NavigationBar registerLink = new NavigationBar(driver);
@@ -78,7 +86,6 @@ public class SignupTests extends NewTestTemplate {
     base.verifyUserLoggedIn(userName);
   }
 
-  @Test(groups = "SignUp_anonCanSignUpWithoutConfirmingVerificationEmail")
   public void anonCanSignUpWithoutConfirmingVerificationEmail() {
     WikiBasePageObject base = new WikiBasePageObject();
     AttachedRegisterPage signUp = base.openSpecialUserSignUpPage(wikiURL);
@@ -96,7 +103,6 @@ public class SignupTests extends NewTestTemplate {
     base.verifyUserLoggedIn(userName);
   }
 
-  @Test(groups = "SignUp_userCanLoginWithoutConfirmingVerificationEmail")
   public void userCanLoginWithoutConfirmingVerificationEmail() {
     WikiBasePageObject base = new WikiBasePageObject();
     AttachedRegisterPage signUp = base.openSpecialUserSignUpPage(wikiURL);
@@ -119,7 +125,6 @@ public class SignupTests extends NewTestTemplate {
     base.verifyUserLoggedIn(userName);
   }
 
-  @Test(groups = "SignUp_anonCanSignUpWithUsernameContainingJapaneseSpecialCharacters")
   @Execute(onWikia = "ja.ja-test")
   public void anonCanSignUpWithUsernameContainingJapaneseSpecialCharacters() {
     WikiBasePageObject base = new WikiBasePageObject();
@@ -137,5 +142,61 @@ public class SignupTests extends NewTestTemplate {
       PageContent.WIKI_SIGN_UP_BIRTHYEAR);
     register.submit();
     base.verifyUserLoggedIn(userName);
+  }
+
+  private void init() {
+    new Navigate().toPageByPath(MercurySubpages.MAIN_PAGE);
+  }
+
+  @InBrowser(browser = Browser.CHROME, emulator = Emulator.GOOGLE_NEXUS_5)
+  public void successfulSignup() {
+    init();
+    signUp(new CreateUser().create()).verifyAvatarAfterSignup();
+  }
+
+  @InBrowser(browser = Browser.CHROME, emulator = Emulator.GOOGLE_NEXUS_5)
+  public void anonCanSignUpWithEmailAlreadyInUse() {
+    init();
+    signUp(new CreateUser().withEmail("qaart001@gmail.com").create()).verifyAvatarAfterSignup();
+  }
+
+  @InBrowser(browser = Browser.CHROME, emulator = Emulator.GOOGLE_NEXUS_5)
+  public void signupErrorUsernameTaken() {
+    init();
+    String userNameTaken = "bekcunning";
+
+    signUp(new CreateUser().withName(userNameTaken).create()).verifyUsernameTakenError();
+  }
+
+  @InBrowser(browser = Browser.CHROME, emulator = Emulator.GOOGLE_NEXUS_5)
+  public void signupErrorBadPassword() {
+    init();
+    String random = "User" + DateTime.now().getMillis();
+
+    signUp(new CreateUser().withName(random).withPass(random).create()).verifyPasswordError();
+  }
+
+  @InBrowser(browser = Browser.CHROME, emulator = Emulator.GOOGLE_NEXUS_5)
+  public void signupErrorTooYoungUser() {
+    init();
+    DateTime wrongBirthDate = new DateTime(2009, 12, 12, 12, 0, 0);
+
+    signUp(new CreateUser().withBirthday(wrongBirthDate).create()).verifyBirthdateError();
+  }
+
+  @InBrowser(browser = Browser.CHROME, emulator = Emulator.GOOGLE_NEXUS_5)
+  @Execute(onWikia = "ja.ja-test")
+  public void japaneseUserSignup() {
+    init();
+    String japanName = "ユーザー" + DateTime.now().getMillis();
+    String japanPAssword = "ユーザザー" + DateTime.now().getMillis();
+
+    signUp(new CreateUser().withName(japanName).withPass(japanPAssword).create())
+      .verifyAvatarAfterSignup();
+  }
+
+  private SignupPageObject signUp(TestUser user) {
+    return new SignupPageObject(driver).openMobileSignupPage().signUp(user.getUserName(),
+      user.getPassword(), user.getEmail(), user.getBirthdate());
   }
 }
